PDFBOX-6131: add cast
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1930890 13f79535-47bb-0310-9956-ffa450edef68

+++ xmpbox/src/test/java/org/apache/xmpbox/xml/DomXmpParserTest.java
@@ -1539,7 +1539,7 @@ public class DomXmpParserTest
         InputStream is = DomXmpParser.class.getResourceAsStream("/org/apache/xmpbox/xml/PDFBOX-6131-RMR6DEEUWZO6IM3A7WKRPX33SZMBTTQZ.xml");
         DomXmpParser xmpParser = new DomXmpParser();
         XMPMetadata xmp = xmpParser.parse(is);
-        assertEquals(1, xmp.getPDFIdentificationSchema().getPart());
+        assertEquals((Integer) 1, xmp.getPDFIdentificationSchema().getPart());
         is.close();
     }
 }
